+++ internal/arm64/bytes_sve2_arm64.go
@@ -127,6 +127,9 @@ func jsonMasksSVE2(dst []byte, b []byte, want uint32)
 func jsonValidTokensSVE2(b []byte, masks []uint64, stk []uint64) int
 
 //go:noescape
+func jsonValidSVE2(b []byte, stk []uint64) int
+
+//go:noescape
 func jsonStage1SVE2(out []uint64, masks []byte, nw int, carr []uint64, res []int64)
 
 //go:noescape